Change default directory permissions
Note that specified value of 0777 is applied with umask,
so by default it will be 0777 - 0022 = 0755.
It allows more flexibility on permissions in a unix-friendly way,
so that if anyone needs to adjust permissions can change process umask.

This is needed for solving permissions issue on DevVM for windows, where
user which runs web application (www-data) and user which shares files
via SMB (vagrant) could create directories, so we use umask 0002 there and
we can achieve writing permissions for group (not only for user).